pyproject.toml
[tool.poetry]
name = "whatson"
version = "0.1.0"
description = "Collect theatre listings in one place"
authors = ["Simon Walker <s.r.walker101@googlemail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.1"
requests = "^2.22.0"
beautifulsoup4 = "^4.8.2"
python-dotenv = "^0.10.3"
psycopg2-binary = "^2.8.4"
lxml = "^4.4.2"
selenium = "^3.141.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
ipython = "^7.11.1"
black = "^19.10b0"
pylint = "^2.4.4"

[tool.poetry.scripts]
whatson-ingest = "whatson.ingest:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
